ProjectUtilsTest moved to the proper (new) project
