+++ dev/tests/hu.elte.txtuml.utils.eclipse.tests/.gitignore
@@ -1,0 +1,3 @@
+bin/
+target/
+
