+++ dev/tests/hu.elte.txtuml.utils.eclipse.tests/META-INF/MANIFEST.MF
@@ -1,0 +1,8 @@
+Manifest-Version: 1.0
+Bundle-ManifestVersion: 2
+Bundle-Name: txtUML Utils Tests
+Bundle-SymbolicName: hu.elte.txtuml.utils.eclipse.tests
+Bundle-Version: 0.5.1.qualifier
+Fragment-Host: hu.elte.txtuml.utils.eclipse
+Bundle-RequiredExecutionEnvironment: JavaSE-1.8
+Require-Bundle: org.junit
