+++ dev/tests/hu.elte.txtuml.utils.eclipse.tests/build.properties
@@ -1,0 +1,3 @@
+source.. = src/
+bin.includes = META-INF/,\
+               .
