+++ dev/tests/hu.elte.txtuml.utils.eclipse.tests/pom.xml
@@ -1,0 +1,14 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<project
+	xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd"
+	xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance">
+	<modelVersion>4.0.0</modelVersion>
+	<parent>
+		<groupId>hu.elte.txtuml</groupId>
+		<artifactId>hu.elte.txtuml.parent</artifactId>
+		<version>0.5.1-SNAPSHOT</version>
+		<relativePath>../../releng/hu.elte.txtuml.parent/pom.xml</relativePath>
+	</parent>
+	<artifactId>hu.elte.txtuml.utils.eclipse.tests</artifactId>
+	<packaging>eclipse-test-plugin</packaging>
+</project>
