+++ dev/tests/hu.elte.txtuml.utils.eclipse.tests/src/hu/elte/txtuml/utils/eclipse/IntegrationTests.java
@@ -1,0 +1,12 @@
+package hu.elte.txtuml.utils.eclipse;
+
+import org.junit.runner.RunWith;
+import org.junit.runners.Suite;
+import org.junit.runners.Suite.SuiteClasses;
+
+//import hu.elte.txtuml.export.papyrus.api.DiagramElementsModifierTest;
+
+@RunWith(Suite.class)
+@SuiteClasses({ProjectUtilsTest.class})
+public class IntegrationTests {
+}
