+++ examples/demo/machine/src/machine1/j/DiagDescWithoutModel.java
@@ -1,0 +1,8 @@
+package machine1.j;
+
+import hu.elte.txtuml.api.layout.ClassDiagram;
+import hu.elte.txtuml.api.layout.Right;
+
+public class DiagDescWithoutModel extends ClassDiagram {
+	class MachineLayout extends Layout {} 
+}
